@@ -1,6 +1,6 @@
 package net.komunan.komunantw.ui.common
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import com.mikepenz.materialdrawer.Drawer
 
 abstract class TWBaseActivity: AppCompatActivity() {
