@@ -1,8 +1,8 @@
 package net.komunan.komunantw.ui.main.home
 
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentManager
-import android.support.v4.app.FragmentPagerAdapter
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentPagerAdapter
 import net.komunan.komunantw.repository.entity.Timeline
 import net.komunan.komunantw.ui.main.home.tab.HomeTabFragment
 
