@@ -1,10 +1,10 @@
 package net.komunan.komunantw.ui.common
 
-import android.arch.lifecycle.LiveData
-import android.arch.lifecycle.MutableLiveData
-import android.arch.lifecycle.Transformations
-import android.arch.lifecycle.ViewModel
 import android.content.Context
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.Transformations
+import androidx.lifecycle.ViewModel
 import net.komunan.komunantw.ReleaseApplication
 
 open class TWBaseViewModel: ViewModel() {
