@@ -1,19 +1,19 @@
 package net.komunan.komunantw.ui.main.sources
 
-import android.arch.lifecycle.ViewModelProviders
 import android.os.Bundle
-import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.lifecycle.ViewModelProviders
 import kotlinx.android.synthetic.main.simple_list_view.*
 import net.komunan.komunantw.R
 import net.komunan.komunantw.observeOnNotNull
+import net.komunan.komunantw.ui.common.TWBaseFragment
 
-class SourcesFragment: Fragment() {
+class SourcesFragment: TWBaseFragment() {
     companion object {
         @JvmStatic
-        fun create(): Fragment = SourcesFragment()
+        fun create() = SourcesFragment()
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
